requirements.txt
alembic==1.6.4
aniso8601==9.0.1
appdirs==1.4.4
click==8.0.1
distlib==0.3.1
filelock==3.0.12
flake8==3.9.2
Flask==2.0.1
Flask-HTTPAuth==4.4.0
Flask-Migrate==3.0.0
Flask-MySQL==1.5.2
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
importlib-metadata==4.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.4
MarkupSafe==2.0.1
mccabe==0.6.1
packaging==20.9
pluggy==0.13.1
py==1.10.0
pycodestyle==2.7.0
pyflakes==2.3.1
PyJWT==2.1.0
PyMySQL==1.0.2
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.15
toml==0.10.2
tox==3.23.1
typing-extensions==3.10.0.0
virtualenv==20.4.7
Werkzeug==2.0.1
zipp==3.4.1
alembic==1.6.4
aniso8601==9.0.1
appdirs==1.4.4
autopep8==1.5.7
click==8.0.1
distlib==0.3.1
filelock==3.0.12
flake8==3.9.2
Flask==2.0.1
Flask-Migrate==3.0.0
Flask-MySQL==1.5.2
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
importlib-metadata==4.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.4
MarkupSafe==2.0.1
mccabe==0.6.1
packaging==20.9
pluggy==0.13.1
py==1.10.0
pycodestyle==2.7.0
pyflakes==2.3.1
PyJWT==2.1.0
PyMySQL==1.0.2
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.15
toml==0.10.2
tox==3.23.1
typing-extensions==3.10.0.0
virtualenv==20.4.7
Werkzeug==2.0.1
zipp==3.4.1
